# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ever Works Website Template API",
    "description": "REST API documentation for the Ever Works website template. This template provides a complete foundation for building modern web applications with authentication, user management, favorites, and content management features.",
    "version": "1.0.0",
    "contact": {
      "name": "Ever Works Template",
      "url": "https://demo.ever.works"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://demo.ever.works",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/api/admin/featured-items": {
      "get": {
        "tags": ["Admin - Featured Items"],
        "summary": "List featured items (Admin)",
        "description": "Returns a paginated list of featured items with filtering options. Requires admin privileges.",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number for pagination",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "example": 1
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            },
            "example": 10
          },
          {
            "name": "active",
            "in": "query",
            "description": "Filter by active status",
            "schema": {
              "type": "boolean"
            },
            "example": true
          }
        ],
        "responses": {
          "200": {
            "description": "Featured items retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminFeaturedItemsResponse"
                },
                "example": {
                  "success": true,
                  "data": [
                    {
                      "id": "featured_123",
                      "itemSlug": "awesome-productivity-tool",
                      "itemName": "Awesome Productivity Tool",
                      "itemIconUrl": "https://cdn.example.com/icons/productivity-tool.png",
                      "itemCategory": "Productivity",
                      "itemDescription": "A powerful tool to boost your productivity",
                      "featuredOrder": 10,
                      "featuredUntil": "2024-12-31T23:59:59.000Z",
                      "isActive": true,
                      "featuredBy": "admin_456",
                      "featuredAt": "2024-01-15T10:30:00.000Z",
                      "createdAt": "2024-01-15T10:30:00.000Z",
                      "updatedAt": "2024-01-15T10:30:00.000Z"
                    }
                  ],
                  "pagination": {
                    "page": 1,
                    "limit": 10,
                    "total": 25,
                    "totalPages": 3,
                    "hasNext": true,
                    "hasPrev": false
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "post": {
        "tags": ["Admin - Featured Items"],
        "summary": "Create featured item (Admin)",
        "description": "Creates a new featured item. Requires admin privileges. Checks for duplicate active featured items.",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFeaturedItemRequest"
              },
              "example": {
                "itemSlug": "awesome-productivity-tool",
                "itemName": "Awesome Productivity Tool",
                "itemIconUrl": "https://cdn.example.com/icons/productivity-tool.png",
                "itemCategory": "Productivity",
                "itemDescription": "A powerful tool to boost your productivity",
                "featuredOrder": 10,
                "featuredUntil": "2024-12-31T23:59:59.000Z"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Featured item created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateFeaturedItemResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "id": "featured_789",
                    "itemSlug": "awesome-productivity-tool",
                    "itemName": "Awesome Productivity Tool",
                    "itemIconUrl": "https://cdn.example.com/icons/productivity-tool.png",
                    "itemCategory": "Productivity",
                    "itemDescription": "A powerful tool to boost your productivity",
                    "featuredOrder": 10,
                    "featuredUntil": "2024-12-31T23:59:59.000Z",
                    "isActive": true,
                    "featuredBy": "admin_456",
                    "featuredAt": "2024-01-15T10:30:00.000Z",
                    "createdAt": "2024-01-15T10:30:00.000Z",
                    "updatedAt": "2024-01-15T10:30:00.000Z"
                  },
                  "message": "Item featured successfully"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation error or duplicate",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "missing_fields": {
                    "summary": "Missing required fields",
                    "value": {
                      "success": false,
                      "error": "Item slug and name are required"
                    }
                  },
                  "already_featured": {
                    "summary": "Item already featured",
                    "value": {
                      "success": false,
                      "error": "Item is already featured"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/api/admin/featured-items/{id}": {
      "get": {
        "tags": ["Admin - Featured Items"],
        "summary": "Get featured item by ID (Admin)",
        "description": "Returns a specific featured item by its ID. Requires admin privileges.",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Unique identifier of the featured item",
            "example": "featured_123"
          }
        ],
        "responses": {
          "200": {
            "description": "Featured item retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetFeaturedItemResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "id": "featured_123",
                    "itemSlug": "awesome-productivity-tool",
                    "itemName": "Awesome Productivity Tool",
                    "itemIconUrl": "https://cdn.example.com/icons/productivity-tool.png",
                    "itemCategory": "Productivity",
                    "itemDescription": "A powerful tool to boost your productivity",
                    "featuredOrder": 10,
                    "featuredUntil": "2024-12-31T23:59:59.000Z",
                    "isActive": true,
                    "featuredBy": "admin_456",
                    "featuredAt": "2024-01-15T10:30:00.000Z",
                    "createdAt": "2024-01-15T10:30:00.000Z",
                    "updatedAt": "2024-01-15T10:30:00.000Z"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Featured item not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": "Featured item not found"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "put": {
        "tags": ["Admin - Featured Items"],
        "summary": "Update featured item (Admin)",
        "description": "Updates a featured item. All fields are optional. Requires admin privileges.",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Unique identifier of the featured item",
            "example": "featured_123"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFeaturedItemRequest"
              },
              "example": {
                "itemName": "Updated Productivity Tool",
                "itemIconUrl": "https://cdn.example.com/icons/updated-tool.png",
                "itemCategory": "Productivity",
                "itemDescription": "An even more powerful tool to boost your productivity",
                "featuredOrder": 15,
                "featuredUntil": "2024-12-31T23:59:59.000Z",
                "isActive": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Featured item updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateFeaturedItemResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "id": "featured_123",
                    "itemSlug": "awesome-productivity-tool",
                    "itemName": "Updated Productivity Tool",
                    "itemIconUrl": "https://cdn.example.com/icons/updated-tool.png",
                    "itemCategory": "Productivity",
                    "itemDescription": "An even more powerful tool to boost your productivity",
                    "featuredOrder": 15,
                    "featuredUntil": "2024-12-31T23:59:59.000Z",
                    "isActive": true,
                    "featuredBy": "admin_456",
                    "featuredAt": "2024-01-15T10:30:00.000Z",
                    "createdAt": "2024-01-15T10:30:00.000Z",
                    "updatedAt": "2024-01-15T12:45:00.000Z"
                  },
                  "message": "Featured item updated successfully"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Featured item not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": "Featured item not found"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "delete": {
        "tags": ["Admin - Featured Items"],
        "summary": "Remove featured item (Admin)",
        "description": "Soft deletes a featured item by setting isActive to false. Requires admin privileges.",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Unique identifier of the featured item",
            "example": "featured_123"
          }
        ],
        "responses": {
          "200": {
            "description": "Featured item removed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteFeaturedItemResponse"
                },
                "example": {
                  "success": true,
                  "message": "Featured item removed successfully"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Featured item not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": "Featured item not found"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token",
        "description": "Session-based authentication using NextAuth.js"
      }
    },
    "schemas": {
...
      "FeaturedItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique featured item identifier"
          },
          "itemSlug": {
            "type": "string",
            "description": "Slug identifier of the featured item"
          },
          "itemName": {
            "type": "string",
            "description": "Name of the featured item"
          },
          "itemIconUrl": {
            "type": "string",
            "format": "uri",
            "nullable": true,
            "description": "URL to the item's icon"
          },
          "itemCategory": {
            "type": "string",
            "nullable": true,
            "description": "Category of the featured item"
          },
          "itemDescription": {
            "type": "string",
            "nullable": true,
            "description": "Description of the featured item"
          },
          "featuredOrder": {
            "type": "integer",
            "description": "Display order for featured items (higher numbers first)"
          },
          "featuredUntil": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "When the featured status expires (null means no expiration)"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the featured item is currently active"
          },
          "featuredBy": {
            "type": "string",
            "description": "ID of the admin user who featured this item"
          },
          "featuredAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the item was featured"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the featured item record was created"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the featured item record was last updated"
          }
        },
        "required": ["id", "itemSlug", "itemName", "featuredOrder", "isActive", "featuredBy", "featuredAt", "createdAt", "updatedAt"]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "details": {
            "type": "string",
            "description": "Additional error details"
          }
        },
        "required": ["success", "error"]
      },
...
      "AdminFeaturedItemsResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FeaturedItem"
            },
            "description": "Array of featured items"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationInfo"
          }
        },
        "required": ["success", "data", "pagination"]
      },
      "CreateFeaturedItemRequest": {
        "type": "object",
        "properties": {
          "itemSlug": {
            "type": "string",
            "minLength": 1,
            "description": "Unique slug identifier of the item to feature",
            "example": "awesome-productivity-tool"
          },
          "itemName": {
            "type": "string",
            "minLength": 1,
            "description": "Name of the item to feature",
            "example": "Awesome Productivity Tool"
          },
          "itemIconUrl": {
            "type": "string",
            "format": "uri",
            "nullable": true,
            "description": "URL to the item's icon",
            "example": "https://cdn.example.com/icons/productivity-tool.png"
          },
          "itemCategory": {
            "type": "string",
            "nullable": true,
            "description": "Category of the item",
            "example": "Productivity"
          },
          "itemDescription": {
            "type": "string",
            "nullable": true,
            "description": "Description of the item",
            "example": "A powerful tool to boost your productivity"
          },
          "featuredOrder": {
            "type": "integer",
            "default": 0,
            "description": "Display order for featured items (higher numbers first)",
            "example": 10
          },
          "featuredUntil": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "When the featured status expires (null means no expiration)",
            "example": "2024-12-31T23:59:59.000Z"
          }
        },
        "required": ["itemSlug", "itemName"]
      },
      "CreateFeaturedItemResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "$ref": "#/components/schemas/FeaturedItem"
          },
          "message": {
            "type": "string",
            "example": "Item featured successfully"
          }
        },
        "required": ["success", "data", "message"]
      },
      "PaginationInfo": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "description": "Current page number",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "description": "Number of items per page",
            "example": 10
          },
          "total": {
            "type": "integer",
            "description": "Total number of items",
            "example": 25
          },
          "totalPages": {
            "type": "integer",
            "description": "Total number of pages",
            "example": 3
          },
          "hasNext": {
            "type": "boolean",
            "description": "Whether there is a next page",
            "example": true
          },
          "hasPrev": {
            "type": "boolean",
            "description": "Whether there is a previous page",
            "example": false
          }
        },
        "required": ["page", "limit", "total", "totalPages", "hasNext", "hasPrev"]
      },
      "GetFeaturedItemResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "$ref": "#/components/schemas/FeaturedItem"
          }
        },
        "required": ["success", "data"]
      },
      "UpdateFeaturedItemRequest": {
        "type": "object",
        "properties": {
          "itemName": {
            "type": "string",
            "minLength": 1,
            "description": "Name of the item (optional)",
            "example": "Updated Productivity Tool"
          },
          "itemIconUrl": {
            "type": "string",
            "format": "uri",
            "nullable": true,
            "description": "URL to the item's icon (optional)",
            "example": "https://cdn.example.com/icons/updated-tool.png"
          },
          "itemCategory": {
            "type": "string",
            "nullable": true,
            "description": "Category of the item (optional)",
            "example": "Productivity"
          },
          "itemDescription": {
            "type": "string",
            "nullable": true,
            "description": "Description of the item (optional)",
            "example": "An even more powerful tool to boost your productivity"
          },
          "featuredOrder": {
            "type": "integer",
            "description": "Display order for featured items (optional)",
            "example": 15
          },
          "featuredUntil": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "When the featured status expires (optional)",
            "example": "2024-12-31T23:59:59.000Z"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the featured item is active (optional)",
            "example": true
          }
        }
      },
      "UpdateFeaturedItemResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "$ref": "#/components/schemas/FeaturedItem"
          },
          "message": {
            "type": "string",
            "example": "Featured item updated successfully"
          }
        },
        "required": ["success", "data", "message"]
      },
      "DeleteFeaturedItemResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Featured item removed successfully"
          }
        },
        "required": ["success", "message"]
      },
...
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Unauthorized - authentication required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "error": "Unauthorized"
            }
          }
        }
      },
      "ServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "error": "Internal server error"
            }
          }
        }
      }
    }
  }
}